never show horizontal scrollbar for branch+commit selection
Somewhat on the fence here. You _can_ scroll sideways with a trackpad or
mouse wheel (or other physical device), but the widget affordances are
gone. But ava's horizontal scrollbar is jumpy af. The scrollbar length
depends on what is currently in view. So as you scroll longer items into
view, the bar size re-renders. Easy to get a re-render every single
frame which looks pretty glitchy. Removing the scroll widget is probably
less accessible, but not any more so than the garbage scrollbars
platform native on macos and windows?