fix branch/commit list/selection reactivity
There should always be a branch and commit selected, if at all possible.
Soons the project is loaded, try to select a branch. Soons a branch is
selected, try to select a commit. If a branch or commit was already
selected, try to maintain that selection. Git fetches should update both
the branch and commit lists, and the current selections (if any) should
be maintained if possible.